# swagger.json
{
	"components": {
...
		"schemas": {
...
			"GetSignupResponse": {
				"properties": {
					"meta": {
						"properties": {},
						"type": "object"
					},
					"data": {
						"properties": {
							"captcha": {
								"properties": {
									"svg": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"svg",
									"id"
								],
								"type": "object"
							}
						},
						"required": [
							"captcha"
						],
						"type": "object"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SignupResponse": {
				"properties": {
					"meta": {
						"properties": {},
						"type": "object"
					},
					"data": {
						"properties": {
							"clientId": {
								"type": "string"
							}
						},
						"required": [
							"clientId"
						],
						"type": "object"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISignupParams": {
				"properties": {
					"captchaId": {
						"type": "string"
					},
					"captchaValue": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"captchaId",
					"captchaValue",
					"firstName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SigninResponse": {
				"properties": {
					"meta": {
						"properties": {},
						"type": "object"
					},
					"data": {
						"properties": {
							"email": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"token": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"lastName",
							"firstName",
							"token"
						],
						"type": "object"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISigninParams": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordResponse": {
				"properties": {
					"meta": {
						"properties": {},
						"type": "object"
					},
					"data": {
						"properties": {},
						"type": "object"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IResetPasswordParams": {
				"properties": {
					"token": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordResponse": {
				"properties": {
					"meta": {
						"properties": {},
						"type": "object"
					},
					"data": {
						"properties": {},
						"type": "object"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IForgotPasswordParams": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyResetPasswordTokenResponse": {
				"properties": {
					"meta": {
						"properties": {},
						"type": "object"
					},
					"data": {
						"properties": {},
						"type": "object"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IVerifyResetPasswordTokenParams": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"clientAccessToken": {
				"type": "apiKey",
				"name": "x-authorization",
				"in": "header"
			},
			"accessToken": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tornado-server",
		"version": "1.1.0",
		"license": {
			"name": "Commercial"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/auth/signup": {
			"get": {
				"operationId": "GetSignupParams",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSignupResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {},
											"data": {
												"captcha": {
													"id": "123123-234234-234234",
													"svg": "<svg></svg>"
												}
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignupResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {},
											"data": {
												"clientId": "123456"
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignupParams"
							}
						}
					}
				}
			}
		},
		"/auth/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SigninResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {},
											"data": {
												"token": "507c7f79bcf86cd7994f6c0e",
												"firstName": "First name",
												"lastName": "Last name",
												"email": "test@test.com"
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISigninParams"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResetPasswordResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {},
											"data": {}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IResetPasswordParams"
							}
						}
					}
				}
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForgotPasswordResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {},
											"data": {}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IForgotPasswordParams"
							}
						}
					}
				}
			}
		},
		"/auth/verify-reset-password-token": {
			"post": {
				"operationId": "VerifyResetPasswordToken",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyResetPasswordTokenResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {},
											"data": {}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerifyResetPasswordTokenParams"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}